move utils from generator to internal
