+++ tools/generator/validate/validation.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 
 	"github.com/Azure/azure-sdk-for-go/tools/generator/autorest/model"
-	"github.com/Azure/azure-sdk-for-go/tools/generator/utils"
+	"github.com/Azure/azure-sdk-for-go/tools/internal/utils"
 )
 
 // MetadataValidateFunc is a function that validates a metadata is legal or not
